% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/featureLayers.R
\name{addEsriHeatmapFeatureLayer}
\alias{addEsriHeatmapFeatureLayer}
\title{Add Esri Heatmap Feature Layer.}
\usage{
addEsriHeatmapFeatureLayer(map, url, radius = 25, gradient = NULL,
  options = featureLayerOptions(), layerId = NULL, group = NULL)
}
\arguments{
\item{radius}{Radius for the heatmap}

\item{gradient}{The gradient}
}
\description{
Add Esri Heatmap Feature Layer.
}
