% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imageMapLayers.R
\name{addEsriImageMapLayer}
\alias{addEsriImageMapLayer}
\title{Render and visualize Image Services from ArcGIS Online and ArcGIS Server.}
\usage{
addEsriImageMapLayer(map, url, options = imageMapLayerOptions(),
  popupFunction = NULL, popupOptions = NULL, layerId = NULL,
  group = NULL)
}
\arguments{
\item{map}{The leaflet map}

\item{url}{URL of the \href{http://resources.arcgis.com/en/help/arcgis-rest-api/#/Image_Service/02r3000000q8000000/}{Image Service}}

\item{options}{options for the image map layer.}

\item{popupFunction}{Uses the provided function to create a popup that will
identify features whenever the map is clicked.
Your function will be passed a GeoJSON FeatureCollection of the features
 at the clicked location and should return the appropriate HTML.
 If you do not want to open the popup when there are no results, return false.}

\item{popupOptions}{See \code{\link[leaflet]{popupOptions}}.}

\item{layerId}{A unique ID for the layer.}

\item{group}{The name of the group this layer should be added to.}
}
\description{
Image Services provide access to raster data through a web service.
}
