% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlDependencies.R
\name{addEsriDependency}
\alias{addEsriDependency}
\alias{addEsriClusterDependency}
\alias{addEsriGeocoderDependency}
\alias{addEsriHeatmapDependency}
\alias{addEsriRenderersDependency}
\title{Adds esri-leaflet dependency to the leaflet widget}
\usage{
addEsriDependency(map)

addEsriClusterDependency(map)

addEsriGeocoderDependency(map)

addEsriHeatmapDependency(map)

addEsriRenderersDependency(map)
}
\arguments{
\item{map}{The leaflet map widget}
}
\description{
Adds esri-leaflet dependency to the leaflet widget

Adds esri-leaflet-clustered-feature-layer dependency to the leaflet widget

Adds esri-leaflet-geocoder dependency to the leaflet widget

Adds esri-leaflet-heatmap dependency to the leaflet widget

Adds esri-leaflet-renderers dependency to the leaflet widget
}
