% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layers.R
\docType{data}
\name{esriBasemapLayers}
\alias{addEsriBaseMapLayer}
\alias{esriBaseMapLabels}
\alias{esriBasemapLayers}
\title{Esri Basemap Layers from \url{https://esri.github.io/esri-leaflet/api-reference/layers/basemap-layer.html}.}
\format{An object of class \code{list} of length 9.}
\usage{
esriBasemapLayers

esriBaseMapLabels

addEsriBaseMapLayer(map, key, autoLabels = FALSE, layerId = NULL,
  group = NULL, options = list())
}
\arguments{
\item{map}{The leaflet map}

\item{key}{ID of the layer}

\item{autoLabels}{whether to show corresponding labels layer}

\item{layerId}{Unique ID for the layer}

\item{group}{The group this layer belongs to.}

\item{options}{Basemap Layer Options.}
}
\description{
Esri Basemap Layers from \url{https://esri.github.io/esri-leaflet/api-reference/layers/basemap-layer.html}.

Esri basemap labels from \url{https://esri.github.io/esri-leaflet/api-reference/layers/basemap-layer.html#optional-labels}.

Adds a ArcGIS Basemap layer
}
\keyword{datasets}

