% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemapLayers.R
\docType{data}
\name{esriBasemapLabels}
\alias{esriBasemapLabels}
\title{Esri basemap labels.}
\format{An object of class \code{list} of length 7.}
\usage{
esriBasemapLabels
}
\description{
Taken from \url{https://esri.github.io/esri-leaflet/api-reference/layers/basemap-layer.html#optional-labels}.
}
\keyword{datasets}
