% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basemapLayers.R
\docType{data}
\name{esriBasemapLayers}
\alias{esriBasemapLayers}
\title{Esri Basemap Layers.}
\format{An object of class \code{list} of length 9.}
\usage{
esriBasemapLayers
}
\description{
BasemapLayer is used to display Esri hosted basemaps and attributes data providers appropriately. The Terms of Use for Esri hosted services apply to all Leaflet applications.
Taken from \url{https://esri.github.io/esri-leaflet/api-reference/layers/basemap-layer.html}.
}
\keyword{datasets}
