% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.esri-package.r
\docType{package}
\name{leaflet.esri}
\alias{leaflet.esri}
\alias{leaflet.esri-package}
\title{leaflet.esri: 'ESRI' bindings for the 'leaflet' Package.}
\description{
An add-on package to the 'leaflet' package, which provides bindings for 'ESRI' services.
This package allows a user to add 'ESRI' provided services such as
'MapService', 'ImageMapService', 'TiledMapService' etc. to a 'leaflet' map.
}
\author{
Bhaskar V. Karambelkar
}
