% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledMapLayers.R
\name{tiledMapLayerOptions}
\alias{tiledMapLayerOptions}
\title{Options for TiledMapLayer.}
\usage{
tiledMapLayerOptions(correctZoomLevels = TRUE, zoomOffsetAllowance = 0.1,
  proxy = NULL, useCors = TRUE, token = NULL, tileOptions = NULL)
}
\arguments{
\item{correctZoomLevels}{Correct Zoom levels.}

\item{zoomOffsetAllowance}{If correctZoomLevels is enabled this controls the amount of tolerance for the difference at each scale level for remapping tile levels.}

\item{proxy}{URL of an \href{https://developers.arcgis.com/javascript/jshelp/ags_proxy.html}{ArcGIS API for JavaScript proxy} or \href{https://github.com/Esri/resource-proxy}{ArcGIS Resource Proxy} to use for proxying requests.}

\item{useCors}{Dictates if the service should use CORS when making GET requests.}

\item{token}{Use this token to authenticate all calls to the service.}

\item{tileOptions}{Other options for tile layer. You can pass \code{\link[leaflet]{tileOptions}()}.}
}
\description{
Options for TiledMapLayer.
}
